+++ source/core/adapters/redux/pulls/reducers.ts
@@ -70,6 +70,7 @@ function listReducer(
 
             return {
                 ...state,
+                type,
                 total,
                 page,
                 perPage,