+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add the loading skeleton.
+
+### Fixed
 - Change javascript into typescript.
 
 ## [0.2.0]